Important bug fix about aveScan arised up using python3 Changing the type of NATIVE_VARS in read_descriptor.py, from set to (ordered) list.
This affects these two lines in aveScan.py
VARLIST=[]
VARLIST.extend(NATIVE_VARS)

In python2, when we convert a set in a list we get always the same order.
In python3, the order is not maintained, every core generates its own VARLIST

When we do
PROCESSES = np.arange(nFiles * nvars)
for ip in PROCESSES[rank::nranks]:
    (ifile, ivar) = divmod(ip,nvars)
    lvar     = VARLIST[ivar]

we must be sure that VARLIST is the same for all the cores.
If not, some files will be processed twice, some others won’t be processed.
That’s why we need an ordered list.